Fixed compilation warnings with GCC

+++ src/hist.cpp
@@ -66,7 +66,7 @@ void Histogram::extractObsScore(const Matrix<float>& R, size_t offset,
                           vector<ScoreHistogram>& histContainer)
 {
         for (size_t j = 0; j < sm.getNumOccCol(); j++) {
-                for (size_t i = 0; i < R.nRows(); i++) {
+                for (size_t i = 0; i < (size_t)R.nRows(); i++) {
                         float thisScore = R(i,j);
                         size_t motifID = motifContainer.getMotifIDAtRow(i);
 
@@ -99,7 +99,7 @@ void Histogram::histThread(const MotifContainer& motifContainer,
         Matrix<float> R(P.nRows(), W);
 
         while (sm.getNextSeqMatrix(seqBatch)) {
-                for (int offset = 0; offset < K; offset++) {
+                for (size_t offset = 0; offset < K; offset++) {
                         SubMatrix<float> subS = sm.getSubMatrix(offset);
                         //R.gemm(P, subS);
                         R.gemm(P, subS, matBlock);
